backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic[email]
yfinance
pandas
numpy
websockets
httpx
python-dotenv
aiosqlite
tensorflow
scikit-learn
ta
apscheduler
